# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Documentação da API",
        "contact": {
            "email": "seu-email@dominio.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/project": {
            "get": {
                "tags": [
                    "public"
                ],
                "summary": "Retornar uma lista de projetos paginada",
                "description": "Retornar uma lista de projetos paginada",
                "operationId": "projectList",
                "responses": {
                    "200": {
                        "description": "Lista de exemplos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "public"
                ],
                "summary": "Inseri um objeto",
                "description": "Cria um novo recurso no banco de dados",
                "operationId": "ProjectStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "type_installation_id",
                                    "uf_id",
                                    "equipament"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "type_installation_id": {
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "uf_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "equipament": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/ProjectEquipament"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exemplo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/project/{id}": {
            "get": {
                "tags": [
                    "public"
                ],
                "summary": "Retornar o projeto e seus respecitvos itens de forma detalhada",
                "description": "Retornar o projeto e seus respecitvos itens de forma detalhada",
                "operationId": "projectDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Projeto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de exemplos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "public"
                ],
                "summary": "Deletar um recurso",
                "description": "Deleta um recurso baseado no ID fornecido",
                "operationId": "projectDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso que será deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso deletado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            }
        },
        "/api/project_equipament/{idProject}/{idEquip}": {
            "put": {
                "tags": [
                    "public"
                ],
                "summary": "Atualizar um objeto",
                "description": "Atualiza um Equipamento a um projeto",
                "operationId": "ProjectEquipamentUpdate",
                "parameters": [
                    {
                        "name": "idProject",
                        "in": "path",
                        "description": "ID do Projeto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idEquip",
                        "in": "path",
                        "description": "ID do Equipamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectEquipament"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exemplo não encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "public"
                ],
                "summary": "Atualizar um objeto",
                "description": "Adiciona um Equipamento a um projeto",
                "operationId": "ProjectEquipamentAdd",
                "parameters": [
                    {
                        "name": "idProject",
                        "in": "path",
                        "description": "ID do Projeto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idEquip",
                        "in": "path",
                        "description": "ID do Equipamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectEquipament"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exemplo não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "public"
                ],
                "summary": "Deletar um recurso",
                "description": "Deleta um recurso baseado no ID fornecido",
                "operationId": "ProjectEquipamentDelete",
                "parameters": [
                    {
                        "name": "idProject",
                        "in": "path",
                        "description": "ID do Projeto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idEquip",
                        "in": "path",
                        "description": "ID do Equipamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso deletado com sucesso"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectEquipament": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Project": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "uf": {
                        "type": "integer",
                        "example": 1
                    },
                    "client": {
                        "type": "integer",
                        "example": 1
                    },
                    "type_installation": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}